Finished with Create API

+++ pet/proxzacom/app/api/posts/[id]/route.ts
@@ -1,4 +1,7 @@
+import prisma from "@/lib/prismadb";
+import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
   try {
@@ -12,6 +15,12 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 }
 
 export async function PUT(req: Request, { params }: { params: { id: string } }) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated!" }, { status: 401 });
+  }
+
   const { title, content, selectCategory, imageUrl, publicId } = await req.json();
   const id = params.id;
 
@@ -35,6 +44,12 @@ export async function PUT(req: Request, { params }: { params: { id: string } }) 
 }
 
 export async function DELETE(req: Request, { params }: { param: { id: string } }) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated!" }, { status: 401 });
+  }
+
   const id = params.id;
   try {
     const post = await prisma.post.delete({ where: { id } });
